updated readme

+++ collatz.py
@@ -11,10 +11,10 @@ print (a)
 while a > 1:
 
     if a % 2 == 0:
-       a =  a/2
+       a =  int(a/2)
        #print (a) for testing
     else:
-        a = (a*3)+1
+        a = int((a*3)+1)
 
     print (a)
